requirements.txt
python-dotenv~=1.2.1
motor~=3.7.1
pymongo~=4.15.3
discord.py~=2.6.4
tabulate~=0.9.0
certifi~=2025.10.5
aiosqlite==0.19.0
pyspellchecker==0.8.1
requests~=2.32.5